(WIP) AutoInvestTransaction.process in progress
Also - refactor some tests
